code splitting refactor
