+++ app/packages/spaces/src/components/EmptySpace.tsx
@@ -1,0 +1,3 @@
+export default function EmptySpace() {
+  return <h1>Empty Space</h1>;
+}
