+++ app/packages/spaces/src/components/index.ts
@@ -1,0 +1,9 @@
+export { default as AddPanelButton } from "./AddPanelButton";
+export { default as AddPanelItem } from "./AddPanelItem";
+export { default as EmptySpace } from "./EmptySpace";
+export { default as Panel } from "./Panel";
+export { default as PanelIcon } from "./PanelIcon";
+export { default as PanelTab } from "./PanelTab";
+export { default as Space } from "./Space";
+export { default as SplitPanelButton } from "./SplitPanelButton";
+export * from "./StyledElements";
