+++ app/packages/spaces/src/enums.ts
@@ -1,0 +1,10 @@
+// an enum for different layouts of a space
+export enum Layout {
+  Vertical = "vertical",
+  Horizontal = "horizontal",
+}
+
+export enum SPACE_TYPES {
+  PANEL_CONTAINER = "panel-container",
+  EMPTY = "empty",
+}
